Organize funcs into src for easier admin
