+++ scripts/main.py
@@ -1,6 +1,6 @@
-from utils import generate_player_ids, model_sign_ons, assign_countries
-from generate_sample_data import generate_sessions
-from load_to_duckdb import connect_to_duckdb
+from src.utils import generate_player_ids, model_sign_ons, assign_countries
+from src.generate_sample_data import generate_sessions
+from src.load_to_duckdb import connect_to_duckdb
 
 def main():
     print("📦 Connecting to DuckDB...")
