+++ src/generate_sample_data.py
@@ -8,9 +8,9 @@ from itertools import islice
 import numpy as np
 import pandas as pd
 
-from load_to_duckdb import write_session_to_disk, stage_session
+from src.load_to_duckdb import write_session_to_disk, stage_session
 
-from utils import SESSION_PATH, HEARTBEAT_INTERVAL, SESSION_MAX_DURATION_SECONDS, \
+from src.utils import SESSION_PATH, HEARTBEAT_INTERVAL, SESSION_MAX_DURATION_SECONDS, \
     MIN_TEAMS, MAX_TEAMS, MIN_PLAYERS_PER_TEAM, MAX_PLAYERS_PER_TEAM
